update account model roi and amount to be type Double, format account amount to look like us currency

+++ I Have More Money Than/I Have More Money Than/Models/Account.swift
@@ -11,6 +11,6 @@ import Foundation
 struct Account: Decodable {
     let id: Int
     let name: String
-    let amount: Float
-    let ROI: Float
+    let amount: Double
+    let ROI: Double
 }
